lib: add additional unescaping on unsexpify
For historycal reasons, xapi's database needed to escape certain characters.
Not everything that needed escape was escaped, and to fix that, future
versions of xapi will escape all single white spaces with '%.'.
This means that serialised sexps will be polluted by lots of '%.'.
This commit fix the most common cases, allowing this tool to keep working
with most databases.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>